Fix expression with different enum types
Arithmetic conversions on operands of different enumeration types
are deprecated in C++20 so we need to explicitly cast the operand
to the compatible type.

Closes  #22505.

+++ include/wx/propgrid/propgridpagestate.h
@@ -125,7 +125,7 @@ wxPG_ITERATOR_MASK_OP_PARENT = wxPG_ITERATOR_FLAGS_ALL,
 
 // Combines all flags needed to iterate through visible properties
 // (ie. hidden properties and children of collapsed parents are skipped).
-wxPG_ITERATE_VISIBLE = wxPG_ITERATE_PROPERTIES |
+wxPG_ITERATE_VISIBLE = static_cast<int>(wxPG_ITERATE_PROPERTIES) |
                        wxPG_PROP_CATEGORY |
                        wxPG_IT_CHILDREN(wxPG_PROP_AGGREGATE),
